// src/types/roomInfo.ts
export type CardColor = 'diamonds' | 'hearts' | 'spades' | 'clubs';

export type CardStatu = 'undistributed' | 'distributed';

export type CardShowFace = 'front' | 'back';

export interface CardType {
  key: string;
  color: CardColor;
  number: number | string;
  /** which face is the card toward to */
  showFace: CardShowFace;
  /** who is the card belong to */
  holder?: string;
  statu: CardStatu;
}

export enum CardColorEnum {
  SPADE = 'spades',
  HEART = 'hearts',
  CLUB = 'clubs',
  DIAMOND = 'diamonds'
}

export type PlayerInfoStatusType = 'disconnect' | 'calling' | 'waiting' | 'fold';

export interface PlayerInfoType {
  name: string;
  position: number;
  status: PlayerInfoStatusType[];
  holdCards: CardType[];
  debt: number;
  calledChips: number;
  holdCent: number;
  blind: number;
  /** whether called in one round */
  roundCalled: boolean;
  /** for heart detection */
  activeTime: number;
}

export interface RoomInfo {
  publicCards: CardType[];
  players: Map<string, PlayerInfoType>;
  buttonIndex: number;
  statu: 'waiting' | 'started' | 'settling';
  currentCallChips: number;
  currentHasChips: number;
  callingSteps: number;
  smallBlind: number;
  bigBlind: number;
  isShortCards: boolean;
}

export interface VictoryInfo {
  getChips: number;
  cardName?: string;
  cards?: CardType[];
}

export interface PlayerCallChipsRes {
  victoryPlayers?: [PlayerInfoType, VictoryInfo][];
  playersCalledRes: [PlayerInfoType, string][];
}

export interface ChatMessageType { 
  key: string | number;
  player: PlayerInfoType; 
  msg: string;
}